Move ICMP code out of firewall.c
