+++ src/firewall.c
@@ -44,17 +44,12 @@
 #include <arpa/inet.h>
 #include <unistd.h>
 #include <sys/uio.h>
-#include <fcntl.h>
 #include <netdb.h>
 #include <sys/time.h>
 
-#include <net/ethernet.h>
-#include <netinet/ip.h>
-#include <netinet/ip_icmp.h>
-#include <netpacket/packet.h>
-
 #include "httpd.h"
 #include "safe.h"
+#include "util.h"
 #include "debug.h"
 #include "conf.h"
 #include "firewall.h"
@@ -64,8 +59,6 @@
 #include "client_list.h"
 #include "commandline.h"
 
-
-int icmp_fd;
 
 static int _fw_deny_raw(const char *, const char *, const int);
 
@@ -202,18 +195,11 @@ arp_get(const char *req_ip)
 int
 fw_init(void)
 {
-    int flags, oneopt = 1, zeroopt = 0;
 	int result = 0;
     int new_fw_state;
 	t_client * client = NULL;
 
-    debug(LOG_INFO, "Creating ICMP socket");
-    if ((icmp_fd = socket (AF_INET, SOCK_RAW, IPPROTO_ICMP)) == -1 ||
-            (flags = fcntl(icmp_fd, F_GETFL, 0)) == -1 ||
-             fcntl(icmp_fd, F_SETFL, flags | O_NONBLOCK) == -1 ||
-             setsockopt(icmp_fd, SOL_SOCKET, SO_RCVBUF, &oneopt, sizeof(oneopt)) ||
-             setsockopt(icmp_fd, SOL_SOCKET, SO_DONTROUTE, &zeroopt, sizeof(zeroopt)) == -1) {
-        debug(LOG_ERR, "Cannot create ICMP raw socket.");
+    if (!init_icmp_socket()) {
         return 0;
     }
 
@@ -261,11 +247,7 @@ fw_set_authservers(void)
 int
 fw_destroy(void)
 {
-    if (icmp_fd != 0) {
-        debug(LOG_INFO, "Closing ICMP socket");
-        close(icmp_fd);
-    }
-
+    close_icmp_socket();
     debug(LOG_INFO, "Removing Firewall rules");
     return iptables_fw_destroy();
 }
@@ -397,74 +379,3 @@ fw_sync_with_authserver(void)
     client_list_destroy(worklist);
 }
 
-/**
- * Ping an IP.
- * @param IP/host as string, will be sent to gethostbyname
- */
-void
-icmp_ping(const char *host)
-{
-	struct sockaddr_in saddr;
-	struct {
-		struct ip ip;
-		struct icmp icmp;
-	} packet;
-	unsigned int i, j;
-	int opt = 2000;
-	unsigned short id = rand16();
-
-	memset(&saddr, 0, sizeof(saddr));
-	saddr.sin_family = AF_INET;
-	inet_aton(host, &saddr.sin_addr);
-#if defined(HAVE_SOCKADDR_SA_LEN)
-	saddr.sin_len = sizeof(struct sockaddr_in);
-#endif
-
-	memset(&packet.icmp, 0, sizeof(packet.icmp));
-	packet.icmp.icmp_type = ICMP_ECHO;
-	packet.icmp.icmp_id = id;
-
-	for (j = 0, i = 0; i < sizeof(struct icmp) / 2; i++)
-		j += ((unsigned short *)&packet.icmp)[i];
-
-	while (j >> 16)
-		j = (j & 0xffff) + (j >> 16);
-
-	packet.icmp.icmp_cksum = (j == 0xffff) ? j : ~j;
-
-	if (setsockopt(icmp_fd, SOL_SOCKET, SO_RCVBUF, &opt, sizeof(opt)) == -1)
-		debug(LOG_ERR, "setsockopt(): %s", strerror(errno));
-
-	if (sendto(icmp_fd, (char *)&packet.icmp, sizeof(struct icmp), 0,
-	           (const struct sockaddr *)&saddr, sizeof(saddr)) == -1)
-		debug(LOG_ERR, "sendto(): %s", strerror(errno));
-
-	opt = 1;
-	if (setsockopt(icmp_fd, SOL_SOCKET, SO_RCVBUF, &opt, sizeof(opt)) == -1)
-		debug(LOG_ERR, "setsockopt(): %s", strerror(errno));
-
-	return;
-}
-
-unsigned short rand16(void) {
-  static int been_seeded = 0;
-
-  if (!been_seeded) {
-    unsigned int seed = 0;
-    struct timeval now;
-
-    /* not a very good seed but what the heck, it needs to be quickly acquired */
-    gettimeofday(&now, NULL);
-    seed = now.tv_sec ^ now.tv_usec ^ (getpid() << 16);
-
-    srand(seed);
-    been_seeded = 1;
-    }
-
-    /* Some rand() implementations have less randomness in low bits
-     * than in high bits, so we only pay attention to the high ones.
-     * But most implementations don't touch the high bit, so we
-     * ignore that one.
-     **/
-      return( (unsigned short) (rand() >> 15) );
-}
